createCollection - metadata JSON
