+++ lib/imageGeneration.ts
@@ -2,7 +2,7 @@ import { experimental_generateImage as generateImage } from "ai";
 import { openai } from "@ai-sdk/openai";
 import { uploadBase64ToArweave } from "@/lib/arweave/uploadBase64ToArweave";
 import createCollection from "@/app/api/in_process/createCollection";
-import { IS_PROD } from "./consts";
+import { uploadMetadataJson } from "./arweave/uploadMetadataJson";
 
 export interface GeneratedImageResponse {
   image: {
@@ -10,10 +10,6 @@ export interface GeneratedImageResponse {
     mimeType: string;
   };
   arweave?: {
-    id: string;
-    url: string;
-  } | null;
-  metadataArweave?: {
     id: string;
     url: string;
   } | null;
@@ -25,11 +21,11 @@ export interface GeneratedImageResponse {
 }
 
 export async function generateAndProcessImage(
-  prompt: string,
+  prompt: string
 ): Promise<GeneratedImageResponse> {
   if (!process.env.OPENAI_API_KEY) {
     throw new Error(
-      "OpenAI API key is missing. Please add it to your environment variables.",
+      "OpenAI API key is missing. Please add it to your environment variables."
     );
   }
 
@@ -58,15 +54,11 @@ export async function generateAndProcessImage(
   // Upload the generated image to Arweave
   let arweaveData = null;
   try {
-    const arweaveResult = await uploadBase64ToArweave(
+    arweaveData = await uploadBase64ToArweave(
       imageData.base64Data,
       imageData.mimeType,
-      `generated-image-${Date.now()}.png`,
+      `generated-image-${Date.now()}.png`
     );
-    arweaveData = {
-      id: arweaveResult.id,
-      url: arweaveResult.url,
-    };
   } catch (arweaveError) {
     console.error("Error uploading to Arweave:", arweaveError);
     // We'll continue and return the image even if Arweave upload fails
@@ -75,23 +67,10 @@ export async function generateAndProcessImage(
   // Upload metadata JSON to Arweave
   let metadataArweave = null;
   try {
-    const metadata = {
-      external_url: `https://inprocess.fun/collect/${IS_PROD ? "base" : "bsep"}:[collectionAddress]`,
-      image: arweaveData ? `ar://${arweaveData.id}` : "",
-      name: prompt,
-    };
-    const metadataBase64 = Buffer.from(JSON.stringify(metadata)).toString(
-      "base64",
-    );
-    const metadataResult = await uploadBase64ToArweave(
-      metadataBase64,
-      "application/json",
-      `metadata-${Date.now()}.json`,
-    );
-    metadataArweave = {
-      id: metadataResult.id,
-      url: metadataResult.url,
-    };
+    metadataArweave = await uploadMetadataJson({
+      arweaveData,
+      prompt,
+    });
   } catch (metadataError) {
     console.error("Error uploading metadata to Arweave:", metadataError);
   }
